refactor: simplify file structure by removing legacy versions
✅ Removed Files:
- Deleted legacy execute-steps.ts (old version)
- Deleted legacy rlhf-system.ts (old version)
- Removed execute-steps-enhanced.ts (renamed)
- Removed rlhf-system-enhanced.ts (renamed)

🔄 Renamed Files:
- execute-steps-enhanced.ts → execute-steps.ts
- rlhf-system-enhanced.ts → rlhf-system.ts

📝 Updated Documentation:
- README.md: Updated all command examples
- RLHF-INTEGRATION.md: Updated file references
- INTEGRATION-STATUS.md: Documented cleanup

🎯 Benefits:
- Single source of truth for each tool
- No more confusion about which version to use
- All enhancements are now default
- Backward compatibility maintained through exports

The system is now simpler and cleaner with all enhancements built-in.
