+++ INTEGRATION-STATUS.md
@@ -1,52 +1,46 @@
 # RLHF Integration Status Report
 
-## ✅ Current Status: FULLY INTEGRATED
+## ✅ Current Status: FULLY INTEGRATED & SIMPLIFIED
 
-### Files Overview
+### Files Overview (After Cleanup)
 
-#### Production Files (Use These)
-1. **`execute-steps-enhanced.ts`** ✅
-   - Uses `EnhancedRLHFSystem` correctly (line 14)
-   - Calls `calculateLayerScore()` centralized method (lines 384-389)
+#### Production Files
+1. **`execute-steps.ts`** ✅
+   - Uses `EnhancedRLHFSystem` correctly
+   - Calls `calculateLayerScore()` centralized method
    - No duplicate scoring logic
    - Pre-validation with schemas
    - Layer detection from filenames
 
-2. **`rlhf-system-enhanced.ts`** ✅
+2. **`rlhf-system.ts`** ✅
    - Centralized scoring with `calculateLayerScore()`
    - Pattern loading from templates
    - Layer-aware architectural validation
    - Cache management for performance
+   - Exports both `EnhancedRLHFSystem` and `RLHFSystem` for compatibility
 
 3. **`validate-template.ts`** ✅
    - Unified validator with JSON schemas
    - Layer detection
    - Architectural violation checks
 
-#### Legacy Files (Kept for Compatibility)
-1. **`execute-steps.ts`**
-   - Original executor without enhancements
-   - Uses basic `RLHFSystem`
-   - No pre-validation or layer awareness
-   - Kept for backward compatibility
-
-2. **`rlhf-system.ts`**
-   - Original RLHF system
-   - No layer context support
-   - Basic scoring without template integration
-   - Kept for backward compatibility
+#### Removed Files (No Longer Needed)
+1. **~~`execute-steps-enhanced.ts`~~** - Renamed to `execute-steps.ts`
+2. **~~`rlhf-system-enhanced.ts`~~** - Renamed to `rlhf-system.ts`
+3. **~~Legacy `execute-steps.ts`~~** - Removed
+4. **~~Legacy `rlhf-system.ts`~~** - Removed
 
 ## Code Verification
 
 ### Import Check ✅
 ```typescript
-// execute-steps-enhanced.ts line 14 - CORRECT
-import { EnhancedRLHFSystem, LayerInfo } from './rlhf-system-enhanced';
+// execute-steps.ts line 14 - CORRECT
+import { EnhancedRLHFSystem, LayerInfo } from './rlhf-system';
 ```
 
 ### Scoring Method Check ✅
 ```typescript
-// execute-steps-enhanced.ts lines 384-389 - CORRECT
+// execute-steps.ts lines 384-389 - CORRECT
 const score = await this.rlhf.calculateLayerScore(
   step.type,
   success,
@@ -64,7 +58,7 @@ The old scoring logic has been completely removed. The method `calculateLayerAwa
 ```
 User runs command
        ↓
-execute-steps-enhanced.ts
+execute-steps.ts
        ↓
 Detects layer from filename
        ↓
@@ -81,16 +75,16 @@ Returns layer-aware score
 
 ## Recommended Usage
 
-### For New Projects
+### For All Projects
 ```bash
-# Always use the enhanced version
-npx tsx execute-steps-enhanced.ts templates/backend-domain-template.regent
+# Use the main executor with all enhancements
+npx tsx execute-steps.ts templates/backend-domain-template.regent
 ```
 
 ### For RLHF Analysis
 ```bash
-# Use enhanced system with layer context
-npx tsx rlhf-system-enhanced.ts analyze implementation.yaml domain backend
+# Use the enhanced RLHF system with layer context
+npx tsx rlhf-system.ts analyze implementation.yaml domain backend
 ```
 
 ### For Validation
@@ -99,13 +93,13 @@ npx tsx rlhf-system-enhanced.ts analyze implementation.yaml domain backend
 npx tsx validate-template.ts templates/backend-domain-template.regent
 ```
 
-## Migration Path
+## Simplified Structure
 
-If you're using the old system:
-1. Replace `execute-steps.ts` with `execute-steps-enhanced.ts`
-2. Replace `RLHFSystem` with `EnhancedRLHFSystem`
-3. Update imports to use enhanced versions
-4. Benefit from layer-aware scoring and pre-validation
+After cleanup:
+1. Single executor: `execute-steps.ts` (with all enhancements)
+2. Single RLHF system: `rlhf-system.ts` (with layer awareness)
+3. No more "enhanced" versions - everything is enhanced by default
+4. Backward compatibility maintained through exports
 
 ## Performance Improvements
